Restructure files to make kafka consumer entry point of app
